// frontend/apps/coze-studio/src/pages/space-skill/types.ts
/*
 * Copyright 2025 ynet-dev Authors
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

export interface SkillInfo {
  skill_id: string;
  space_id?: string;
  name?: string;
  description?: string;
  prompt?: string;
  icon_uri?: string;
  type?: string;
  product_id?: string;
  files?: Record<string, string>;
  asset_summary?: {
    count?: number;
    image_paths?: string[];
  };
  metadata?: {
    category?: string;
    version?: string;
    tags?: string[];
    platforms?: string[];
  };
  version?: number;
  publish_scope?: number;
  published_version?: number;
  created_at?: number;
  updated_at?: number;
}